Refactor storage utilities into modular class-based internals
- split storage utils into types, serializers, providers and store files
- implement createStorage() with internal StorageStore class
- keep public API unchanged via storage.ts barrel export
- update storage docs to match current provider semantics
